Time: 1 ms (97.04%), Space: 42.6 MB (33.36%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -1,11 +1,11 @@
 class Solution {
 
     public int reverse(int x) {
-        return rev(x, 0);
+        return rev((long)x, 0);
     }
     
-    private int rev(int x, int res) {
-        if (x == 0) return res;
+    private int rev(long x, long res) {
+        if (x == 0) return (int)res;
         if (res > Integer.MAX_VALUE / 10 || res < Integer.MIN_VALUE / 10) return 0;
         return rev(x / 10, res * 10 + x % 10);
     }
